day 4/4주차과제자료/example/requirements.txt
functions-framework==3.5.0
google-cloud-bigquery>=3.5.0
google-auth>=2.32.0
Flask>=2.2.5
